requirements.txt
web3 == 6.20.2
mysql.connector
py-solc-x
PySimpleGUI
pyautogui
psutil
#qui dentro dobbiamo inserire tutto quello che c'è da installare per far funzionare l'applicazione
#ad esempio se usiamo pyGui dobbiamo mettere anche quella etc.
